LayoutTests/media/video-controller-currentTime-rate-expected.txt


Test that setting playbackRate on media controller modifies current time

EVENT(canplaythrough)
EXPECTED (0 == '0') OK
END OF TEST
